feat: support the configuration keep_duplicate_comments

+++ notifier/gitlab/notify.go
@@ -40,7 +40,9 @@ func (g *NotifyService) Notify(ctx context.Context, body string) (exit int, err 
 	value := template.GetValue()
 
 	if cfg.MR.IsNumber() {
-		g.client.Comment.DeleteDuplicates(value.Title)
+		if !cfg.KeepDuplicateComments {
+			g.client.Comment.DeleteDuplicates(value.Title)
+		}
 	}
 
 	_, isApply := parser.(*terraform.ApplyParser)
